Improve functionality of calculator
Ensure operations are performed for only two number pairs. This is so
that the result is displayed before a new operation is done.
Move the calculation to a function for reusability.
Make sure values are reset after equal sign is pressed before new numbers
are displayed. This prevents buggy (concatenation of values displayed)
results.